transcription_app/pyproject.toml
[tool.poetry]
name = "transcription-app"
version = "0.1.0"
description = ""
authors = ["Celina <celinalouissaint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Flask = "^2.2"
SpeechRecognition = "^3.8"
openai = "^1.64.0"
dotenv = "^0.9.9"
pydub = "^0.25.1"
pyannote-audio = "^3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
